Added detail about the post-processing portion. It is being expanded in order to assist post-processing improve.
The project table should only have one state value in my opinion. The way you can represent different needs of a project is by marking each page
(eg: if a project is missing an image for a project, the page can be marked as NEEDS_IMAGE, the rest are marked as HAS_IMAGE, and the project is marked as NEEDS_IMAGE)

Last of all, the "advanced" XML markup should occur after posting of the project. Our PP pool has been growing larger and larger, we need to get the books available at all costs ASAP.