Drop the graph_archive / ignore_archive_graph code.
1) It's unreliable. It archives the graph for month X only if someone
requests it in the first 4 hours of month X+1, which often doesn't
happen. Moreover, if the request happens before the stats have been
recorded for the last day of the month, the graph will be incomplete.

2) It's impolite. It hijacks the user's request, and sends them
back to stats_central.

3) It wastes disk space. About 5k per graph, times ~10 per month,
times 12 months per year, is 0.6 Meg per year. Not a huge amount,
but it's still a waste, since we keep the raw data for all time,
and could regenerate the graphs on demand (if there was any demand).

4) We don't even use it.  The site doesn't link to (or even hint
of) the archive, so the graphs are effectively inaccessible. And
there are no tasks in the Task Center requesting that old graphs be
provided, so even if we did provide access, there's no indication
that people would use it.