Remove the "Project Files" field from the Create Project page (and the code to process it).
For one thing, unpacking an uploaded zip file directly into the project
directory is a fairly serious security hazard. More generally, this
mechanism of adding files to a project is inconsistent in several ways
with that of the project page's Add/Replace feature.

These problems could be resolved by changing the "Project Files" feature
to reuse most of the add_files.php code; however:
-- That would require some effort.
-- It's not clear where the user-interaction of add_files.php would fit
   into the user-interaction of the Create Project page.
-- In the forum topic, 'Use of "Project Files" on Create Project page?',
   none of the posters said they'd be bothered to see this feature go.
-- It's redundant, given the Add/Replace button.