Overhaul. (Sorry for the big delta.)
The intent of this delta was to achieve cleaner and more obvious
control-flow and data-flow, with no user-visible changes.
However, some changes happened along the way...

---

If 'Create Project from Uber Project' fails because of a problem
with the uberproject id, there's no point showing the edit form.
(It's a system error that needs to be fixed before the user can continue.)

Similarly with creating from a MARC record.

----

Formerly, if you hit "Save and Quit" or "Save and Go To Project",
and there were problems with your data, you got error messages,
but you were told to hit 'Back' to fix the problems,
whereupon you could no longer see the error messages.

Instead, if there are problems, these two cases behave more like
"Save and Preview", i.e. show the errors *and* the edit-form
on the same page.

---

Formerly, if you did a SaveAndPreview on an existing project,
and the save raised errors, the form would be pre-filled with values
from the db, and any changes you had made would be forgotten.

Instead, pre-fill the form with values from the preceding form.