When a PM uses the    "Move all P1.unavailable projects on this page to P1.waiting" on a set of projects, they generally end up with the same modifieddate.
Formerly, this would mean that their relative order in the release queue
would be unpredictable, which is particularly undesireable if they are
the parts of a multi-project set.

Add a secondary sort on nameofwork, which (a) makes their relative order
deterministic, and (b) will generally put multi-project sets in order,
assuming they have names of the form "Blah blah blah, part X of Y",
and don't have more than 9 parts.