"X active users out of Y total users in the past twenty-four hours."
1)
The wording was poor: "in the past twenty-four hours" only applies to X,
but the syntax suggests that it applies to Y or to both.
("There were Y total users in the past 24 hours? What does that mean?")

That could have been fixed by saying
    "Out of Y total users, X were active in the past twenty-ofour hours."
or
    "X active users in the last 24 hours out of a total of Y proofreaders."
(see http://www.pgdp.net/phpBB2/viewtopic.php?p=131922#131922)
However...

2)
For Y, we used the number of users with a positive page count in the
"Entry Level" round (P1 on pgdp.net), which was a somewhat clunky
carryover from the R1/R2 days, where we used the number of users with
a positive (R1+R2) page count. The exclusion of zero-page users was
probably an attempt to give a more accurate indication of the pool
of potentially active users, instead of the actual total number of
registered users.
(http://www.pgdp.net/phpBB2/viewtopic.php?p=131912#131912)

In http://www.pgdp.net/phpBB2/viewtopic.php?p=170302#170302, garweyne
suggested that Y should be the number of users with a positive count
in either P1 or R*. Granted, this would have removed the discontinuity
between the old definition and the new, but at the expense of adding
another pgdp.net-specific bit of code.

Instead, if the intent of Y is to give a sense of the pool of potentially
active users, then it seems to me that the purpose would be served by
giving the number of active users over longer timescales.

So that's what I've done.