Add image to spellcheck in the standard interface (Task #965).
Formerly (in std), the spellcheck page would occupy the whole 'proofframe',
which would replace the imageframe+textframe frameset
with a text-only page, and thus "hide" the page image.

Now (in std), the spellcheck doc is directed to just 'textframe',
so that 'imageframe' remains, showing the image. (Note that when
leaving spellcheck mode, we have to be careful about what replaces
the spellcheck doc.)

(Original implementation by cpeel, tweaked by jmdyck.)