Rather than allowing apostrophe anywhere in a word, we now only allow it if it's surrounded by non-apostrophe word characters.
(I.e., we disallow it from initial & final positions,
and we disallow multiple adjacent apostrophes.
The text can have such things, but those apostrophes
are not considered part of a word.)

This should make WordCheck's idea of a word
match aspell's idea of a word in English.
Other languages are not so fortunate.