Move the initialization of JS variables cRef and markRef from initializeStuff() in dp_scroll.js to the onLoad code generated by ctrl_frame.php.
When set correctly, cRef and markRef refer to elements within the
'menuframe' (i.e. toolbox) part of the proofing interface frame-tree.
But initializeStuff() is called from onLoad code in the 'proofframe'
(text + image) part of the frame-tree. Thus, when it's invoked,
it's possible that the toolbox has not yet been loaded/constructed.
In such a case, initializeStuff() would (formerly) set cRef and markRef
to null, causing most of the toolbox functionality to fail.  (This was
a particular problem for the new quiz interface under development.)

Instead, initializeStuff() should not reference anything in menuframe,
and cRef and markRef should be set only after the elements they refer to
are certain to exist.

[The Proofing Interface Toolbox (comprising code in ctrl_frame.php,
toolbox.inc, and dp_proof.js) should be more of a stand-alone unit,
and this move takes us a step closer in that direction.]