Ensure that the correct links are shown for type-in projects that may have no OCR text but will have P1 or later text.
Added comment in show_project_stealth_scannos.php such that if the code is
changed to allow analysis using pages other than OCR as the base, that
edit_project_word_lists.php needs to be updated to show the link.

Finally, tweak variable naming to be more consistent with what they are
used for.