For the "projects you've helped format and/or proof" listing...
If project A was merged into project B,
check whether B has been archived:
if it has, don't display info pertaining to A.

When we used the page_events table, we didn't need this check,
because when B is archived, we remove A-related rows
(as well as B-related rows) from page_events,
so there are no rows in page_events that lead to archived projects.
(So, theoretically, we didn't need the
    AND projects.archived = 0
restriction in the main query either.)

But the archiving process doesn't remove any information from
user_project_info, so we need to check for archived-ness.