Add condition     WHERE email LIKE '%@%.%' to the SQL query.
Say someone's email address is
    someone@hotmail.com
but they mistyped it, using a comma instead of a dot. Then
    SUBSTRING_INDEX(email,'.',-1)
will return their whole address, exposing it in the graph's legend.
To avoid such cases, we do a basic sanity check on the email address.

(At pgdp.net, we've been lucky in that such cases tend to be one-of-a-kind,
and so very low on the list, and so truncated from the graph's legend.)