Remove the use of mb_encode_numericentity from software.php. We've decided not to rely on it, because it's in mbstrings (a not-installed-by-default PHP extension).
Getting software.php to use UTF-8 as the HTML character encoding is not an
option at the moment either, because strings in the theme could rely on
ISO-Latin-1.

As a result, we need to change the native language name strings in
iso_639_loc_list.inc from UTF-8 to HTML entities which we can display
directly on the page. Since the array is only used in a few limited places
(all where HTML is outputted), this isn't a problem, but add a comment
in case someone else runs across this problem.