Another quiz-related patch, courtesy of acunning40.
Allow the forbiddentext tests to take either a string or array of strings
as the search argument. If an array is supplied, fail the test if any
of the items in the array are found.

Use this to fold several tests into one in the code for a number of
quizzes with no user-visible change to the quiz behaviour.