onsider the Task Summary, and the various escapings it undergoes as it meanders its way through browsers, HTTP, PHP, and MySQL...
When a user fills in the Add Task form, the Summary data is "pure" (not
escaped), and we can assume that any escaping it undergoes when packed
into an HTTP request is undone at the server either by Apache or PHP.
But because we have magic_quotes_gpc set, when we pull the data out of
$_POST['task_summary'] (at line 294), it is backslash-escaped (as if by
addslashes).  We then push the data through htmlspecialchars() (with the
default $quote_style, ENT_COMPAT) before we INSERT it into the db. (We
also push it through another call to addslashes, but that's immediately
undone by the MySQL parser.) So the data stored in the database is
effectively:
    htmlspecialchars(addslashes(pure_data), ENT_COMPAT)

When we show the Summary data in the (multiple tasks) Listing page (line
580 in function ShowTasks), or in the (single task) Details page (line
683 in function TaskDetails), we just extract it from the database
and apply stripslashes, which (kind of) undoes the addslashes, leaving
    htmlspecialchars(pure_data, ENT_COMPAT)
which we then insert into the HTML output. The ampersand-escaping
ensures that any HTML-special characters (ampersand, less-than,
greater-than, or double-quote) that the user typed will appear to
the user as typed.

But when we show the Summary data in the Edit Task page (line 612 in
function TaskForm), there's an extra call to htmlspecialchars, which
causes HTML-special characters to appear to the user as their
ampersand-escaped versions (&amp;, &lt;, &gt;, or &quot;, respectively).
We fix this bug by removing the extra call. However, this uncovers
further problems:

 -- In the Listing and Details cases, the Summary data was inserted into
    the HTML output as general element content, so the $quote_style
    passed to htmlspecialchars was unimportant. But in the Edit Task
    case, the Summary data is inserted as the value of an attribute
    (of an <input> element), so the $quote_style becomes important.
    As noted above, it was ENT_COMPAT, in which single-quotes are not
    ampersand-escaped (but this was being masked by the bogus call to
    htmlspecialchars(_, ENT_QUOTES) that we just removed). Thus, in
    TaskForm, we either have to "manually" ampersand-escape just those
    single-quotes, or else delimit the attribute value with
    double-quotes (at line 620). (For now, we pick the latter, as it's
    simpler.) Either way, what the user sees is the unescaped special
    characters, and that's what goes into the form submission.

 -- But now, when we pull that data out of $_POST (line 328), it will
    only have magic_quote_gpc's implicit slash-escaping, and no
    ampersand-escaping. So we have to push it through htmlspecialchars
    before putting it into the UPDATE command, so that the data in the
    database continues to be
	htmlspecialchars(addslashes(pure_data), ENT_COMPAT)
    as expected.

These three changes mostly fix this bug. However, there's still the
nagging problem that if, in the past, someone went into Edit Task and
entered HTML-special characters in the Summary field, those characters
would have been saved to the database (and then subsequently inserted
into Listing and Details pages) without ampersand-escaping. (It looks
like this has actually happened on pgdp.net.) So we may need to clean
up some of the task_summary data in the 'tasks' table.

(Really, I think it would be a lot simpler if we stored the "pure data"
in the database. But that would definitely require tweaking the 'tasks'
table.)