Change how the table-chunks are sorted.
Formerly, we'd sort the chunks simply by the text of the chunk.
Since the table-name was the first point of variation between
chunk-texts, this would basically sort the chunks by table-name.
However, there was a wrinkle when one table-name was a prefix of
another (e.g., phpbb_posts and phpbb_posts_text, or tasks and
tasks_comments). In such cases, because the table-names are quoted,
the sort would be comparing (e.g.)
    ... `phpbb_posts` ...
    ... `phpbb_posts_text` ...
and so the collation order of the backtick came into play. In these
examples, underscore comes before backtick, so the chunks would
sort as
    ... `phpbb_posts_text` ...
    ... `phpbb_posts` ...
which is not what we expect when sorting by table-name.

This commit fixes that.