show_project_listing(): Drop the code that allowed users #1342 & #12417 to see people's email addresses.
In the CVS history, this code first appeared on 2003-10-25
in rev 1.8 of pinc/show_projects_in_state.inc with the comment:
	Updated to show email if user is a site admin or PPVer
	or the PM page if a regular user
So perhaps, at that time, 1342 and 12417 were the only PPVers,
and there wasn't a generic way to identify them.

Anyway, it probably doesn't make sense for those users to continue
to have that privilege, and it certainly doesn't make sense for the
code to identify them in this way. (Other sites that use this code
will eventually have their own users 1342 and 12417!)