Typo correction.
NOTE: The existing comments in this script indicate that it was intended
as a one-time-only method to populate a db table with data scraped from
specially marked Project Comments.

If this is truly the case, this script should probably be moved out of
CVS and into noncvs or the Attic.