This is Step 1/6 of an effort to redesign the search part of the Statistics Central page (stats/stats_central.php).
The table at the top of Statistics Central (containing Member and Team
Search) has been reformated so that it isn't output by echo() statements
from PHP anymore. Instead, we close the PHP block above it and reopen it
below, and just use <?php echo ... ?> where we need PHP variables.
This the first step to improving the readability somewhat.