This is Step 4/6 of an effort to redesign the search part of the Statistics Central page (stats/stats_central.php).
The two rows of the table at the top of Statistics Central have been
merged so that there are just two side-by-side cells left in the table,
one for member search and one for team search.
At the same time, all the "align" attributes have been removed from the
table as we have decided there's no benefit in having things centered.
Both cells are now left-aligned.