This is Step 5/6 of an effort to redesign the search part of the Statistics Central page (stats/stats_central.php). It also includes changes in stats/members/mbr_list.php and stats/teams/tlist.php.
This step implements task 1140 (also see the related forum thread at:
http://www.pgdp.net/phpBB2/viewtopic.php?t=39812).

In stats_central.php, two checkboxes (labeled "Exact match") have been
added below the two text entry fields in the table at the top.

In mbr_list.php and tlist.php, the code now tests whether the relevant
checkbox was checked when the form was submitted. If so, a user/team
with the name as input into the box is searched, instead of any user/team
that contains the entered string. This prevents the user seeing several
pages of matches when they know the exact name of the user/team they are
looking for.
This feature will be especially useful for people who have to grant
access to other members.