This is Step 6/6 of an effort to redesign the search part of the Statistics Central page (stats/stats_central.php).
The table at the top of Statistics Central has been changed to use CSS
instead of HTML attributes. The margin that was added this way makes the
<br> below the table superfluous, so it has been deleted.
Also, the left margin of the text entry fields and the checkboxes has
been set to zero to make things line up in various browsers.