This change fixes task 1165: the PM "Fix" page (tools/project_manager/handle_bad_page.php) eating the first blank line of the text being fixed.
In other words:
added a \n      [pronounce \n as "newline"]
stops the text box from munching
proofers' dear blank lines

The HTML 4.01 spec does not explicitly state how browsers are to treat
the content of <textarea>s starting with newlines. The example (here:
http://www.w3.org/TR/html401/interact/forms.html#h-17.7) is said to
"contain[] two lines of text initially", and it has a newline directly
after the opening tag (that presumably isn't to be displayed as a blank
line).
Safari, FF2, FF3.5, Opera and Camino on my Mac and IE6 (in a VirtualBox
with Windows XP installed) agree with that example and ignore the first
newline in a textarea. So adding a newline there doesn't mess with pages
_not_ starting with a new paragraph.

According to research by acunning40, the same fix was implemented (most
probably sometime in 2004) when the proofing interface used to eat blank
lines.