Include a comment on why we're doing the prep_numeric_keys_for_multisort.
Remove processing for the $instance_total array as it isn't defined at this
point in the script and was inserted due to a copy/paste error from around
line 141.

Remove setting of freqCutoff - likely a result of a copy+change error from
show_good_word_suggestions.php when the stealth scannos file was created.