search_and_list_tasks(): In the SQL query, change     task_foo LIKE '%...%' to     INSTR(task_foo, '...')
When certain characters (percent, underscore, and backslash) appear in
the value of the right-hand operand of LIKE, they are treated specially.
We make use of this by beginning and ending that value with percents.
However, when a special character appears *between* those two percents,
because either:
(a) the user typed it in the text box, or
(b) (backslash only) it was introduced by addslashes,
then the special treatment does not accord with what the user expects.

So instead, use INSTR(), which does not treat any characters specially,
yielding behaviour that probably matches the user's expectations better.