For the transition from     {last_round}->project_complete_state to     PROJ_POST_FIRST_AVAILABLE, add     checkedoutby='' to the settings_template.
This fixes a bug that Jana noticed.

(Originally, that extra setting would have been redundant,
as you could only get a transition to PROJ_POST_FIRST_AVAILABLE
if checkedoutby was already ''. Conversely, checkedout != ''
would cause a transition to PROJ_POST_FIRST_CHECKED_OUT.
But in rev 1.30 of pinc/Project.inc (2008/07/26),
lmpryor added code to project_get_auto_PPer()
to enforce additional requirements of the would-be PPer,
resulting in the possibility of a transition to
PROJ_POST_FIRST_AVAILABLE even though checkedout != ''.
That code was installed 2008-10-10:
http://www.pgdp.net/phpBB2/viewtopic.php?p=495079#495079
so that's how long the bug has existed in the wild.)