TaskDetails(): In Row 5 (Close Task. Me Too!), fix a bunch of well-formedness errors in HTML output (now that they're obvious):
Hoist the <tr> outside the if-stmt,
so that it appears even if neither if-condition is true.

The 'Close Task' form goes *inside* the <td>
(move <form> start-tag down)
and *outside* the 'taskplain' table
(move </form> end-tag down).

Move "</small>" inside second if-arm,
so it only appears if "<small>" does.

Move "</td>" inside both if-arms,
so it only appears if "<td>" does.