TaskDetails(): BUG:     $result = mysql_query(...);     while ($row = mysql_fetch_assoc($result)) {         $result = mysql_query(...); The inner setting of $result clobbers the outer setting, which is still in use as the loop-controller.
The only reason it's managed to work till now is that
there are even more settings of $result within the loop body,
and it happens that the last one of them is always exhausted,
so the loop-controlling call to mysql_fetch_assoc()
returns FALSE for *that* resource,
instead of returning FALSE for the outer (clobbered) resource
(which could only have had one task-row in it anyway),
but the outcome is the same: the loop exits.

(But if you refactor away those later uses of $result,
the last use is the one shown above,
and that resource *isn't* exhausted,
so the loop-controlling call to mysql_fetch_assoc()
returns a usersettings-row, and the loop runs forever,
with copious errors.)

This bug goes all the way back to rev 1.1.

Fix it by renaming the outer uses to '$res'.