OrderBy(): When $_GET['orderby'] is unset, or when it's set but differs from $orderby_var, what is 'direction' in the return value?
Up to rev 1.104, it was 'desc'.
Since rev 1.105, it's been $_GET['direction'] (or, if that's unset, 'asc').

This change in behavior may have been unintended,
since wasn't called out in 1.105's commit message.
I think the former behaviour was better, so reinstate that.

(It might be even better to allow each column to specify its default direction.)