SearchParams_get_url_query_string(): Fix quasi-bug...
Consider this case:
You submit a search+list request without a setting for one (or more) of
the search-parameters (e.g., task_type). You get a listing of tasks, then
click on a column-header to sort the list.

The code in clause_all_or_match() says that the original search will
treat the unset parameter as if it had been set to 999, i.e. the SQL
query's WHERE clause will effectively have no restriction on the
parameter's corresponding column. (This is analogous to what
SearchParams_echo_controls() does: if you don't touch a select-box
in the search form, then that select-box will supply '{param}=999'
to the form submission.)

But when you click on a column-header, you're following a link that's
been generated by SearchParams_get_url_query_string(). Formerly, that
function would treat the unset parameter as if it had been set to a
specific (non-999) value
    (e.g., '1' => 'Bug Report' for task_type)
which would then go into the resulting URL
    (e.g., ...&task_type=1&...)
and thus when you request that URL, SearchParams_get_sql_condition()
would construct an SQL condition that was constrained on that parameter.
That is, clicking on the column-header didn't just sort the tasks,
it also performed a more constrained search than the original, probably
resulting in a smaller set of tasks. This behaviour would be very
surprising.

So, change SearchParams_get_url_query_string() so that, like the other
SearchParams functions, it treats unset parameters as if they were '999'.
(Note that this already happened for task_assignee in rev 1.106.)

This is only a quasi-bug because the code doesn't actually *generate*
search+list requests with unset search parameters. However, consider that
users may have bookmarked particular search+list URLs; if we later add
another search parameter, then those bookmarked URLs (which of course
don't supply a value for that new parameter) should continue to perform
the same search as they did before, i.e. unconstrained on that parameter.
And now they will.