Under isset($_POST['newtask']):
According to the online PHP manual,
    mysql_query("INSERT ...");
    mysql_query("SELECT ...");
    ... mysql_insert_id() ...
shouldn't work. That is, mysql_insert_id() is supposed to return 0
"if the previous query does not generate an AUTO_INCREMENT value".
In practice, mysql_insert_id() appears to ignore the existence of
the SELECT query, and return the AUTO_INCREMENT value generated by
the INSERT query, which is, of course, what we want.

However, we shouldn't rely on that. The PHP manual even has a note:
"Because mysql_insert_id() acts on the last performed query, be sure
to call mysql_insert_id() immediately after the query that generates
the value." So change the code to:
    mysql_query("INSERT ...");
    $task_id = mysql_insert_id();
    mysql_query("SELECT ...");
    ... $task_id ...