Change the value of $default_order_by from:     "ORDER BY date_edited DESC, task_severity ASC, task_type ASC" to just:     "ORDER BY date_edited DESC" (I.e., delete the sub-orders.)
Note that 'date_edited' records time down to the second,
so it's unlikely that two tasks have the same date_edited.
Thus, "ORDER BY date_edited DESC" is likely to completely order
all tasks, and sub-orderings will be unused.

And in the unlikely event that two tasks *do* have the same date_edited,
there doesn't seem to be a compelling reason to sub-order them by
task_severity and task_type.

(Initially, the default order-by was:
    task_severity ASC, date_opened ASC, task_type ASC
Then aldarondo changed it to:
    date_edited DESC, task_severity ASC, task_type ASC
on 2004-01-25 in rev 1.6.)