Drop the $order_by parameter from     list_all_open_tasks()     search_and_list_tasks()     select_and_list_tasks() and instead move the code that sets $order_by (by consulting $_GET) from the top-level 'else' clause (line 750-ish) down into select_and_list_tasks().
Now, for the top-level 'else' clause, it's clear that this is
equivalent -- the same code gets executed, just later/deeper.

For all other calls to
    list_all_open_tasks()
    and
    search_and_list_tasks(),
it's not strictly equivalent. Those calls passed $default_order_by,
without consulting $_GET. But note that all those calls occur when
we're processing a POST request, so $_GET will (normally) be empty,
and the moved code will set:
    $req_direction = 'desc';
    $req_order = 'date_edited';
    $order_by = "ORDER BY date_edited desc";
which is the same as $default_order_by
(modulo the capitalization of 'desc').

(It's possible for a user, using mechanisms outside the site code,
to manufacture a POST request for which $_GET is non-empty.
In such a case, the new code may come up with a different $order_by
than the old code would have, but there's no harm in that.)