Alter the parameter conventions of POST requests.
    edit_task=ID          -> action=show_editing_form & task_id=ID
    reopen_task=ID        -> action=reopen            & task_id=ID
    newtask               -> action=create_or_edit
    search_task           -> action=search
    close_task            -> action=close
    new_comment=ID        -> action=add_comment       & task_id=ID
    new_relatedtask=ID    -> action=add_related_task  & task_id=ID
    new_relatedposting=ID -> action=add_related_topic & task_id=ID
    meToo=ID              -> action=add_metoo         & task_id=ID

(The main purpose is to separate the concerns of "what is to be done"
from "which task to do it to" (where applicable). These are now
consistently conveyed by 'action' and 'task_id' respectively.)

We don't have to add code to convert old-style POST requests to new-style
(as we did with GET requests), because people can't/shouldn't have
old-style POST requests bookmarked or otherwise saved.

We do have to worry about the transition period around installation. E.g.:
-- Shortly before install, someone gets a form and starts filling it out.
-- The install happens.
-- They submit the filled-in form.
In such a case, the new code would receive an old-style request.
However, I plan to install this change during well-advertised
site downtime, so I can assume that people won't get a form
shortly before downtime and expect to be able to submit it after.