In the date() format string     "jS of F, Y" the letters 'o' and 'f' are meant to be passed through as-is, yielding     "6th of December, 2011" for example.
In versions of PHP up to 5.1.0, they are, but in 5.1.0+,
'o' is ascribed a meaning (ISO-8601 year number),
so the same format string yields
    "6th 2011f December, 2011"
for example.

So when pgdp.net upgraded from 4.4.2 to 5.2.4, we began to see the
latter behavior, reported here:
http://www.pgdp.net/phpBB2/viewtopic.php?p=809399#809399

Fix:

To prevent the 'o' from being expanded, we escape it with a backslash.
(See Example #2 at http://php.net/manual/en/function.date.php)

And it's only a matter of time until 'f' is ascribed a meaning,
so escape it too.

But \f in a double-quoted string means form feed (in PHP 5.2.5+),
so put the string in single-quotes.