html_header(): In constructing the <link> that references .../main.css, replace $relPath with $code_url/pinc.
Normally, the two are equivalent,
just relative and absolute paths (respectively)
to the same point in the URL hierarchy.

However, theme.inc is also (on pgdp.net) included by 404.php
(not currently in the codebase), which apache invokes to build
the response to requests for nonexistent resources.
Since the latter can 'be' anywhere in the URL hierarchy,
and since the response is 'located' at the requested URL,
the physical location of the 404.php script can not (in general)
correspond to the "logical" location of the response that it generates.

And since $relPath must reflect the physical location of the script,
and a relative url in <link> is resolved against the logical location
of the response in which it appears, the lack of correspondence means
that its use here generates a link that is incorrect for most
invocations of 404.php. (Which then causes a 404 during the processing
of the original 404.)

Thanks to Donovan for spotting the problem.