Regarding the (user-level) blurb in the project history saying which fields were changed by an 'edit' event...
Formerly,
save_to_db() in tools/project_manager/editproject.php
    would construct the blurb (via get_changed_fields())
    and put it in the event's details1, and then
do_history() in project.php
    would simply echo it.

Instead, save_to_db() now just constructs a list of the
(internal) names of the fields that have changed,
puts that in details1, and do_history() expands that
into a user-level blurb.

This makes it easier to localize the blurb for the user who's
viewing the project history.