Values of $f_f end up as the value of the 'font-family' property in the 'style' attribute of the proofreading <textarea>. (See PPage::echo_proofing_textarea().)
Therefore, change "Monospaced" to "monospace",
so that it's a valid CSS <generic-family>.

The 'd' was definitely a mistake.

The 'M' doesn't appear to be wrong according to the CSS spec,
but anecdotal evidence indicates that some browsers
behave better with 'm'.

(This should fix the problem described in the Details of Task 1484,
though not the problem described in AndiPersti's comment on 2012-01-08.)

[On behalf of andipersti, Donovan, and me.]