This is a grouped change which functionally separates the definitions of user_teams.createdby and user_teams.owner. Formerly, .createdby merely stored the username of the .owner (as u_id, at team creation) and it was in mixed use for display and determination of team ownership.
With this patchset, we are now able to define user_teams.createdby solely
as the username of the original creator of the team, and user_teams.owner
as the sole determinator of who is the leader of the team. This will allow
us (in a future patchset) to preserve the team originator and be able to
assign a new leader for cases where the originator is long-absent and
an active volunteer wishes to take on the team leader role.