Task #1364: Make most task list columns not wrap on resize
The task center task lists behave much better on window resizes if
most of the columns are forbidden to wrap onto a second line. This
prevents the row height increasing too quickly as you make the page
narrower.

To make this scale better, remove the width:50% from the summary
column, and make the 'Site Administrator Request' task type string
shorter.