Initial commits.
Currently, the "popup help" buttons on the User Preferences and Team pages
take their content from very old static HTML files.

These include files now provide that information in a form which is easier
to maintain and allows the popups to be localized.

Calls to the static html files can now be replaced with calls to these files.