Discontinue quiz-page-specific $links_out -- *always* use the links that qp_echo_solved_html() generates.
There were only 4 quiz-pages that defined $links_out...

In one case (p_basicx_1), $links_out provided 2 links:
    "Take another 'extra practice' quiz" and
    "Back to proofreading quiz home"
These are the same links that qp_echo_solved_html() would generate, except
that for the first, the link-text would simply be "Next step of quiz".
While the $links_out text is better, it's not really enough to justify
maintaining $links_out.

In 3 cases (p_fraktur, p_aeoe_2, p_thorn), $links_out provided just one link:
    "Back to proofreading quiz home"

 -- For p_fraktur, this is the same as what qp_echo_solved_html() would
    generate, so there's no point to defining $links_out.

 -- For p_aeoe_2 and p_thorn, it omits the "Next step of tutorial|quiz" links
    that qp_echo_solved_html() would generate. I suppose this is because
    the "Other Specialized Quizzes and Tutorials" is seen more as a collection
    of unrelated 1- and 2-page quizzes than as a single coherent quiz.
    So, e.g., after the "Thorn" page, "Fraktur" isn't really the "next step".
    On the other hand, they *are* structured as a single quiz on the quiz home
    page, and you can get a "Quiz passed" if you pass all 4 pages. So it seems
    impolite to not provide a link to step through the pages.

From the point of view of software design, it's better if a qd file doesn't
need to know which (if any) quiz-page is "next" (or even that there is a
concept of "next"), or what particular URL is needed to get to it.