qp_choose_solution():
Formerly, in the hypothetical case where
    count($solutions) > 1 and !isset($criteria),
we would skip the loop and simply use the last entry in $solutions,
always ignoring the other solutions.

Instead, assert that this case cannot happen --
if there are multiple solutions, there must be selection criteria,
and there must be the correct number of them.

Move some comments too.