diff():
Formerly, we would swap $arr1 and $arr2 if necessary
so that count($arr1) >= count($arr2).
This had the benefit of simplifying the logic in the body of the loop,
but had the disadvantage that we no longer knew
which was the user-text and which the solution-text,
resulting in rather vague difference-descriptions.

Instead, don't do the swap.
Make the necessary changes to the loop.