qp_describe_difference():
Formerly, when we exhausted one text without finding any differences,
we would find the first non-blank line in the other text's "excess",
and then output "This is the first differing line:" followed by that line
(not indicating whether it was a line from the user's text or the solution text).

To me, this seems like not the clearest way to describe the difference.

Instead, tell the user that they need to delete/add some lines (as appropriate)
at the end of their text.