title_case(): Change str[i] to str.charAt(i) because accessing a string as if it were an array of characters is non-standard and not entirely portable (though widely enough supported to have survived here for quite a while).
This appears to have been the cause of the "Abc" button not working for
some cases involving Internet Explorer, as reported in
    http://www.pgdp.net/c/tasks.php?action=show&task_id=1533
and
    http://www.pgdp.net/phpBB2/viewtopic.php?t=53136

The change to title_case() was proposed by dkretz, but when cpeel
implemented it in his sandbox, it didn't seem to fix the problem.
(See http://www.pgdp.net/phpBB2/viewtopic.php?p=899872#899872
and following.) I suspect that, although the change *did* allow
affected browsers to get through title_case()'s main loop, they would
still raise an error in the call to lc_common(), due to the use of
    words[0][0]
which also treats a string (words[0]) as an array of characters.

While it would work to change that expression to
    words[0].charAt(0)
it's simpler for lc_common() to just leave words[0] with
whatever casing it had initially (which should be Titlecase,
because that's what title_case() does).