maybe_mail(): Fix bug: PHP's string-concatenation operator is dot, not plus.
The plus operator coerces its operands to numbers and adds them.
Since the strings at the point in question weren't numeric,
they coerced to zero, and $additional_headers was assigned the value 0,
losing any useful headers it had formerly contained.

(Apparently the mail() function is smart enough to not actually include "0"
as a mail header, so it included it at the start of the body instead.)