In calls to show_preference()     for Name and Email in echo_general_tab(),     and     for Current Profile in echo_proofreading_tab() ...
Formerly, we would pass array('', '') to the $extras parameter,
causing _show_textfield() to generate an <input> tag with the attribute size='',
and no text following the tag.

The 'no text following' is fine, but size='' meant that, for these 3 textfields,
users were getting the browser's default textfield width.
HTML5 specifies this to be 20 characters,
so I'm guessing that's what most modern browsers do,
but HTML4 didn't specify it at all,
so there might be browsers (or browser-modes) out there with different defaults.

There's no reason to let it default.

So instead:
   For Name and Current Profile, make it explicitly 20.
   For Email, set it to 26.
       (see http://www.pgdp.net/phpBB2/viewtopic.php?p=963245#963245,
       which is what prompted this change)