ReleaseRestrictor::update_for_released_project(): Fix quasi-bug: Increment $this->this_round_pms for the released project.
It's only a quasi-bug because approves_project() rejects
any project whose PM matches that of a released-this-time project.
So any project that approves_project() formerly failed to reject
because we weren't properly maintaining $this->this_round_pms
would have been rejected anyway based on $this->fresh_released_pms.

Still, that's somewhat fragile, and there's virtually no cost to
incrementing $this->this_round_pms, so better to do it.